# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Froglet Node API
  version: 1.1.0
  description: >
    HTTP API for a Froglet node. Froglet is a service-execution framework; the
    node exposes two distinct ports:

    - **Provider port** (default 8080) — public requester endpoints plus
      Bearer-authenticated local provider-control endpoints. Do not expose a
      provider-control token to requesters.

    - **Runtime port** (default 8081) — Bearer-authenticated endpoints used by
      the local operator to create deals via the marketplace, manage the wallet,
      and launch workloads.

    Both ports also expose the shared `/health`, `/v1/node/capabilities`, and
    `/v1/node/identity` endpoints.
servers:
  - url: http://localhost:8080
    description: Provider port (public requester routes plus authenticated local control routes)
  - url: http://localhost:8081
    description: Runtime port (local operator, Bearer auth required)
...
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Static Bearer token configured via `FROGLET_RUNTIME_AUTH_TOKEN`.
        Pass as `Authorization: Bearer <token>` on all runtime-port endpoints
        that require authentication.
    ProviderControlBearerAuth:
      type: http
      scheme: bearer
      description: >
        Static Bearer token configured via `FROGLET_PROVIDER_CONTROL_AUTH_TOKEN`.
        Pass as `Authorization: Bearer <token>` on provider-control endpoints.
...
  schemas:
    Error:
      type: object
      description: Standard error envelope returned on 4xx and 5xx responses.
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: deal not found
        details:
          type: string
          description: Additional diagnostic context, present on some errors.
        deal_id:
          type: string
          description: >
            Echoed deal ID, present when the error is scoped to a specific deal.
        service_id:
          type: string
          description: Echoed service ID on publication lifecycle errors.
        revision_hash:
          type: string
          description: Echoed immutable revision hash on revision-scoped errors.
        lifecycle_status:
          type: string
          description: Current lifecycle classification when it explains a conflict.
        remediation:
          type: string
          description: Safe next action when the server can provide one.
        actual_bytes:
          type: integer
          description: Observed request size on bounded publication errors.
        max_bytes:
          type: integer
          description: Enforced byte limit paired with `actual_bytes`.
...
    SignedArtifact:
      type: object
      description: >
        Froglet signed artifact envelope. Every on-chain document is wrapped in
        this envelope. The `payload` field carries the type-specific data.
        `hash` is the SHA-256 of the canonical signing bytes; `signature` is
        the hex BIP340 Schnorr signature over those same bytes.
      required:
        - artifact_type
        - schema_version
        - signer
        - created_at
        - payload_hash
        - payload
        - signature
      properties:
        artifact_type:
          type: string
          description: >
            Kind discriminator. One of `descriptor`, `offer`, `quote`, `deal`,
            `receipt`, `curated_list`, `confidential_profile`,
            `confidential_session`, `invoice_bundle`.
          example: deal
        schema_version:
          type: string
          description: Always `froglet/v1` for the current protocol revision.
          example: froglet/v1
        signer:
          type: string
          description: Hex-encoded 32-byte x-only secp256k1 public key of the signing node.
          example: "abcd1234..."
        created_at:
          type: integer
          format: int64
          description: Unix timestamp (seconds) when the artifact was signed.
        payload_hash:
          type: string
          description: SHA-256 hex digest of the canonical JSON payload.
        hash:
          type: string
          description: >
            SHA-256 hex digest of the canonical signing bytes. Serves as the
            global artifact identifier.
        payload:
          type: object
          description: Type-specific payload. See the concrete payload schemas below.
          additionalProperties: true
        signature:
          type: string
          description: BIP340 Schnorr signature (hex) over the canonical signing bytes.
    DescriptorPayload:
      type: object
      description: >
        Payload carried inside a `descriptor` SignedArtifact. Describes the
        provider's identity, transport endpoints, and declared capabilities.
      required:
        - provider_id
        - protocol_version
        - capabilities
      properties:
        provider_id:
          type: string
          description: Stable identifier for this provider (hex public key).
        descriptor_seq:
          type: integer
          format: int64
          description: Monotonically increasing descriptor sequence number.
        protocol_version:
          type: string
          example: froglet/v1
        expires_at:
          type: integer
          format: int64
          description: Optional Unix timestamp after which this descriptor should be refreshed.
        linked_identities:
          type: array
          description: External identities linked to this provider (e.g. Nostr pubkey).
          items:
            $ref: '#/components/schemas/LinkedIdentity'
        transport_endpoints:
          type: array
          description: Reachable transport endpoints for this provider.
          items:
            $ref: '#/components/schemas/TransportEndpoint'
        capabilities:
          type: object
          description: Declared service and execution capabilities.
          properties:
            service_kinds:
              type: array
              items:
                type: string
              example: ["compute.execution.v1"]
            execution_runtimes:
              type: array
              items:
                type: string
              example: ["wasm", "python", "container"]
            max_concurrent_deals:
              type: integer
              description: Maximum concurrent deals the provider will accept. Absent means no declared limit.
    LinkedIdentity:
      type: object
      description: Cross-identity link (e.g. a Nostr public key attesting this provider).
      required:
        - identity_kind
        - identity
        - scope
        - created_at
        - signature_algorithm
        - linked_signature
      properties:
        identity_kind:
          type: string
          example: nostr
        identity:
          type: string
          description: The external identity (e.g. Nostr hex pubkey).
        scope:
          type: array
          items:
            type: string
          description: Scopes granted to this linked identity.
        created_at:
          type: integer
          format: int64
        expires_at:
          type: integer
          format: int64
        signature_algorithm:
          type: string
          example: secp256k1_schnorr_bip340
        linked_signature:
          type: string
    TransportEndpoint:
      type: object
      description: A reachable network endpoint for this provider.
      required:
        - transport
        - uri
        - priority
      properties:
        transport:
          type: string
          example: clearnet
        uri:
          type: string
          example: https://provider.example
        created_at:
          type: integer
          format: int64
        expires_at:
          type: integer
          format: int64
        priority:
          type: integer
          format: int32
          description: Lower value means higher preference.
        features:
          type: array
          items:
            type: string
    OfferPayload:
      type: object
      description: >
        Payload inside an `offer` SignedArtifact. Describes a priced execution
        offer that requesters can reference when creating quotes.
      required:
        - provider_id
        - offer_id
        - descriptor_hash
        - offer_kind
        - settlement_method
        - quote_ttl_secs
        - execution_profile
        - price_schedule
      properties:
        provider_id:
          type: string
        offer_id:
          type: string
          example: service.echo
        descriptor_hash:
          type: string
        expires_at:
          type: integer
          format: int64
        offer_kind:
          type: string
          example: compute.execution.v1
        settlement_method:
          type: string
          example: lightning
        quote_ttl_secs:
          type: integer
          format: int64
        execution_profile:
          type: object
          description: Resource limits and runtime constraints for this offer.
          properties:
            runtime:
              type: string
              example: wasm
            package_kind:
              type: string
            contract_version:
              type: string
            access_handles:
              type: array
              items:
                type: string
            abi_version:
              type: string
            capabilities:
              type: array
              items:
                type: string
            max_input_bytes:
              type: integer
            max_runtime_ms:
              type: integer
              format: int64
            max_memory_bytes:
              type: integer
            max_output_bytes:
              type: integer
            fuel_limit:
              type: integer
              format: int64
        price_schedule:
          type: object
          properties:
            base_fee_msat:
              type: integer
              format: int64
              description: Fixed fee in millisatoshis regardless of outcome.
            success_fee_msat:
              type: integer
              format: int64
              description: Fee in millisatoshis charged only on successful execution.
        terms_hash:
          type: string
        confidential_profile_hash:
          type: string
...
    ExecutionMount:
      type: object
      additionalProperties: false
      required:
        - handle
        - kind
        - read_only
      properties:
        handle:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-z0-9_]+$'
        kind:
          type: string
          enum: [postgres, sqlite, object_store, redis, s3]
          description: >
            Runtime records may contain legacy s3. Capability grants remain
            exact to the record's kind and are not interchangeable aliases.
        read_only:
          type: boolean
        binding:
          type: string
          description: Provider-resolved binding; omitted from public catalog views.
...
    ProviderControlOfferRecord:
      type: object
      description: Provider-control view of a published offer plus local binding metadata.
      properties:
        publication_state:
          type: string
          enum: [active, hidden]
        service_id:
          type: string
        project_id:
          type: string
        source_kind:
          type: string
        runtime:
          type: string
        package_kind:
          type: string
        entrypoint_kind:
          type: string
        entrypoint:
          type: string
        contract_version:
          type: string
        mounts:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionMount'
        capabilities:
          type: array
          description: Provider-required capability strings such as `compute.gpu`.
          items:
            type: string
        mode:
          type: string
        summary:
          type: string
        module_hash:
          type: string
        binding_hash:
          type: string
        starter:
          type: string
          description: Compact JSON example input for LLM/client guidance; not stronger than `input_schema`.
        input_schema:
          type: object
          additionalProperties: true
        output_schema:
          type: object
          additionalProperties: true
        offer:
          allOf:
            - $ref: '#/components/schemas/SignedArtifact'
            - type: object
              properties:
                payload:
                  $ref: '#/components/schemas/OfferPayload'
    ProviderControlMutationResponse:
      type: object
      required:
        - request_id
        - status
        - summary
        - artifacts
        - evidence
        - offer
      properties:
        request_id:
          type: string
        status:
          type: string
        failure_kind:
          type: string
        summary:
          type: string
        artifacts:
          type: array
          items:
            type: object
            required: [kind, hash]
            properties:
              kind:
                type: string
              hash:
                type: string
        evidence:
          $ref: '#/components/schemas/ProviderControlEvidence'
        offer:
          $ref: '#/components/schemas/ProviderControlOfferRecord'
    ProviderControlEvidence:
      type: object
      additionalProperties: false
      required:
        - provider_id
        - descriptor_hash
        - offer_hash
        - offer_id
        - lifecycle_status
      properties:
        provider_id:
          type: string
          pattern: '^[0-9a-f]{64}$'
        descriptor_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        offer_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        offer_id:
          type: string
        lifecycle_status:
          type: string
          enum: [active, paused, legacy_unverified]
          description: >
            Local immutable-revision lifecycle state after publication.
            `legacy_unverified` is returned only for compatibility offers that
            have no lifecycle-eligible verified revision.
        service_id:
          type: string
        local_verification:
          $ref: '#/components/schemas/LocalVerificationEvidence'
        publication_revision:
          $ref: '#/components/schemas/SignedPublicationRevisionV1'
        activation_token:
          type: string
          pattern: '^[0-9a-f]{64}$'
          description: >
            Opaque identifier for the exact active lifecycle instance created
            by this verified publish. Present for lifecycle-eligible publishes
            and omitted for legacy unverified compatibility offers.
...
paths:
...
  /v1/provider/descriptor:
    get:
      operationId: getProviderDescriptor
      summary: Provider descriptor
      description: >
        Returns the signed descriptor artifact for this provider. The descriptor
        contains the provider's identity, transport endpoints, and capability
        advertisement. Requesters should verify the artifact signature before
        trusting any of its fields.
      tags: [Catalog]
      security: []
      responses:
        '200':
          description: Signed descriptor artifact.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SignedArtifact'
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/DescriptorPayload'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/provider/offers:
    get:
      operationId: listProviderOffers
      summary: List active offers
      description: >
        Returns all signed offer artifacts currently published by this provider.
        Only `active` offers are returned by default. Each offer specifies a
        priced execution profile that requesters can reference in quote requests.
      tags: [Catalog]
      security: []
      responses:
        '200':
          description: Array of signed offer artifacts.
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/SignedArtifact'
                    - type: object
                      properties:
                        payload:
                          $ref: '#/components/schemas/OfferPayload'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/provider/files/abort:
    post:
      summary: Abort active provider file responses
      description: Requires provider control bearer authentication. Already received bytes cannot be revoked. Usage is preserved.
      security:
        - ProviderControlBearerAuth: []
      responses:
        '200': {description: Abort generation advanced}
        '401': {description: Provider control authentication required}
...
  /v1/provider/artifacts/publish:
    post:
      operationId: publishProviderArtifact
      summary: Publish an artifact-backed provider service
      description: >
        Provider-control endpoint for publishing a locally built artifact as a
        named service through `froglet.publication-intent.v1`. When a private
        verification fixture is supplied, the provider executes it locally
        before persistence and returns `LocalVerificationEvidence` plus a
        `froglet.publication-revision.v1` signed statement. These are
        application-layer publication records, not Kernel artifacts.
        `starter`, `input_schema`, and `output_schema` are client guidance
        metadata and do not replace handler-side validation.
      tags: [Provider Control]
      security:
        - ProviderControlBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderControlPublishArtifactRequest'
      responses:
        '201':
          description: >
            Published offer metadata and signed artifact references. Local
            verification and publication revision evidence are present only
            when a verification fixture ran successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderControlMutationResponse'
        '400':
          description: Invalid publication request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid provider-control bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Verification input or expected output exceeds a publication limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Local verification was unsafe, unavailable, failed, or mismatched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Local verification execution capacity is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...